+++ pkg/analyzer/lib/diagnostic/diagnostic.dart
@@ -126,7 +126,7 @@ class Diagnostic {
   int get offset => problemMessage.offset;
 
   Severity get severity {
-    switch (errorCode.errorSeverity) {
+    switch (errorCode.severity) {
       case DiagnosticSeverity.ERROR:
         return Severity.error;
       case DiagnosticSeverity.WARNING:
@@ -134,7 +134,7 @@ class Diagnostic {
       case DiagnosticSeverity.INFO:
         return Severity.info;
       default:
-        throw StateError('Invalid error severity: ${errorCode.errorSeverity}');
+        throw StateError('Invalid error severity: ${errorCode.severity}');
     }
   }