+++ pkg/analyzer/lib/src/dart/error/syntactic_errors.g.dart
@@ -2072,7 +2072,7 @@ class ParserErrorCode extends DiagnosticCode {
        );
 
   @override
-  DiagnosticSeverity get errorSeverity => DiagnosticSeverity.ERROR;
+  DiagnosticSeverity get severity => DiagnosticSeverity.ERROR;
 
   @override
   DiagnosticType get type => DiagnosticType.SYNTACTIC_ERROR;