+++ pkg/analyzer/lib/src/pubspec/pubspec_warning_code.g.dart
@@ -262,7 +262,7 @@ class PubspecWarningCode extends DiagnosticCode {
        );
 
   @override
-  DiagnosticSeverity get errorSeverity => DiagnosticSeverity.WARNING;
+  DiagnosticSeverity get severity => DiagnosticSeverity.WARNING;
 
   @override
   DiagnosticType get type => DiagnosticType.STATIC_WARNING;