+++ pkg/linter/tool/benchmark.dart
@@ -218,7 +218,7 @@ int _maxSeverity(List<DiagnosticInfo> infos) {
   var filteredErrors = infos.expand((i) => i.diagnostics);
   return filteredErrors.fold(
     0,
-    (value, e) => math.max(value, e.errorCode.errorSeverity.ordinal),
+    (value, e) => math.max(value, e.errorCode.severity.ordinal),
   );
 }